Move more implementations into submodules.
